add float32 type instead float64

+++ internal/models/balance.go
@@ -7,5 +7,5 @@ type Withdraw struct {
 	UserID       int       `db:"user_id"`
 	OrderNum     string    `db:"number" json:"order"`
 	DateWithdraw time.Time `db:"date_withdraw" json:"processed_at"`
-	Sum          float64   `db:"sum" json:"sum"`
+	Sum          float32   `db:"sum" json:"sum"`
 }
